misc(finance): move FractionUnit implementations in sub modules
