+++ platform/packages/finance/src/percent/mod.rs
@@ -1,7 +1,4 @@
-use std::ops::{Div, Rem};
-
 use bound::BoundPercent;
-use gcd::Gcd;
 
 use crate::{
     error::Error,
@@ -9,10 +6,10 @@ use crate::{
     fractionable::{CommonDoublePrimitive, Fractionable, FractionableLegacy, IntoMax},
     ratio::{Ratio, SimpleFraction},
     rational::{Rational, RationalLegacy},
-    zero::Zero,
 };
 
 pub mod bound;
+mod fraction;
 mod fractionable;
 
 pub type Units = u32;
@@ -20,32 +17,6 @@ pub type Percent100 = BoundPercent<{ Percent::HUNDRED.units() }>;
 pub type Percent = BoundPercent<{ Units::MAX }>;
 
 // TODO revisit it's usage after removing FractionLegacy<Units> for Percent100
-impl FractionUnit for Units {
-    type Times = Self;
-
-    fn gcd<U>(self, other: U) -> Self::Times
-    where
-        U: FractionUnit<Times = Self::Times>,
-    {
-        Gcd::gcd(self, other.to_primitive())
-    }
-
-    fn scale_down(self, scale: Self::Times) -> Self {
-        debug_assert_ne!(scale, Self::Times::ZERO);
-
-        self.div(scale)
-    }
-
-    fn modulo(self, scale: Self::Times) -> Self::Times {
-        debug_assert_ne!(scale, Self::Times::ZERO);
-
-        self.rem(scale)
-    }
-
-    fn to_primitive(self) -> Self::Times {
-        self
-    }
-}
 
 impl Percent100 {
     pub const fn complement(self) -> Self {
